+++ CHANGES.md
@@ -1,3 +1,11 @@
+#### 0.0.4 / 2022-05-09
+* Added theme setting.
+* Added message setting.
+* Added Dismiss text setting.
+* Added learn more text and link settings.
+* Added placeholders to fields to show default setting or example setting.
+* Improved default settings resilience
+
 #### 0.0.3 / 2022-05-09
 * Added position setting
 * Added defaults when no value set